requirements.txt
speechRecognition==3.10.0
pyaudio
pyttsx3==2.90
openai==0.27.4
streamlit
langchain==0.0.208
unstructured
tiktoken
faiss.cpu
bokeh==2.4.2
streamlit-bokeh-events==0.1.2
gTTs==2.3.1
